+++ src/powerquery-formatter/index.ts
@@ -1,7 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from "./format";
-export * from "./serialize";
 export * from "./formatV2";
 export * from "./serializeV2";
