+++ src/powerquery-formatter/passes/index.ts
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export { tryTraverseSerializeParameter } from "./serializeParameter";
 export { tryTraverseSerializeParameterV2 } from "./serializeParameterV2";
-export { tryTraverseComment } from "./comment";
 export { tryTraverseCommentV2 } from "./commentV2";
 export * from "./commonTypes";
