+++ src/test/smallPrograms.ts
@@ -2,9 +2,9 @@
 // Licensed under the MIT license.
 
 import "mocha";
-import { compare, expectFormat } from "./common";
+import { compareV2, DefaultFormatSettings2, expectFormatV2 } from "./common";
 
-describe(`small programs`, () => {
+describe(`small programs V2`, () => {
     it(`fastPow`, async () => {
         const expected: string = `
 // taken from: https://en.wikipedia.org/wiki/Exponentiation_by_squaring
@@ -18,28 +18,28 @@ describe(`small programs`, () => {
 //     else if n is odd then return x * exp_by_squaring(x * x, (n - 1) / 2);
 let
     isEven = (x as number) => Number.Mod(x, 2) = 0,
-    pow =
-        (x as number, p as number) =>
-            if p = 0 then
-                1
-            else if p < 0 then
-                error "negative power not supported"
-            else
-                x * @pow(x, p - 1),
-    fastPow =
-        (x as number, p as number) =>
-            if p = 0 then
-                1
-            else if p < 0 then
-                error "negative power not supported"
-            else if isEven(p) then
-                @fastPow(x * x, p / 2)
-            else
-                x * @fastPow(x * x, (p - 1) / 2)
+    pow = (x as number, p as number) =>
+        if p = 0 then
+            1
+        else if p < 0 then
+            error "negative power not supported"
+        else
+            x * @pow(x, p - 1),
+    fastPow = (x as number, p as number) =>
+        if p = 0 then
+            1
+        else if p < 0 then
+            error "negative power not supported"
+        else if isEven(p) then
+            @fastPow(x * x, p / 2)
+        else
+            x * @fastPow(x * x, (p - 1) / 2)
 in
-    fastPow(2, 8)`;
+    fastPow(2, 8)
+`;
 
-        const actual: string = await expectFormat(`
+        const actual: string = await expectFormatV2(
+            `
 // taken from: https://en.wikipedia.org/wiki/Exponentiation_by_squaring
 // removed negative powers, sure to have bugs
 //
@@ -70,8 +70,13 @@ let
             else
                 x * @fastPow(x * x, (p - 1) / 2)
 in
-    fastPow(2, 8)`);
+    fastPow(2, 8)`,
+            {
+                ...DefaultFormatSettings2,
+                maxWidth: 120,
+            },
+        );
 
-        compare(expected, actual);
+        compareV2(expected, actual);
     });
 });
